src/pyproject.toml
[tool.poetry]
name = "PyInventory"
version = "0.4.2"
description = "Web based management to catalog things including state and location etc. using Python/Django."
authors = ["JensDiemer <git@jensdiemer.de>"]
packages = [
    { include = "inventory" },
    { include = "inventory_project" },
]
keywords=['inventory','django']
classifiers = [
    # http://pypi.python.org/pypi?%3Aaction=list_classifiers
    "Development Status :: 4 - Beta",
    #"Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3 :: Only",
    'Framework :: Django',
    "Topic :: Database :: Front-Ends",
    "Topic :: Documentation",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Internet :: WWW/HTTP :: Site Management",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
    "Operating System :: OS Independent",
]
include = ['AUTHORS', 'LICENSE', 'README.creole']

# Will be generated from README.creole with: 'poetry run update_rst_readme'
readme='README.rst'

[tool.poetry.dependencies]
python = ">=3.7,<4.0.0"
colorama = "*"  # Console colors under windows: https://pypi.org/project/colorama/
colorlog = "*"  # https://github.com/borntyping/python-colorlog
uWSGI = "*"
#
# https://www.djangoproject.com/download/#supported-versions
# v2.2 LTS - extended support until April 2022
django = "2.2.*"
django-processinfo = "*"  # https://github.com/jedie/django-processinfo/
django-debug-toolbar = "*"  # http://django-debug-toolbar.readthedocs.io/en/stable/changes.html
django-import-export = "*"  # https://github.com/django-import-export/django-import-export
django-dbbackup = "*"  # https://github.com/django-dbbackup/django-dbbackup
django-tools = "*"  # https://github.com/jedie/django-tools/
django-reversion-compare = "*"  # https://github.com/jedie/django-reversion-compare/
django-ckeditor = "*"  # https://github.com/django-ckeditor/django-ckeditor

# We can't install v5 until https://github.com/python-poetry/poetry/issues/3320 is fixed!
bx_py_utils = "!=5"  # https://github.com/boxine/bx_py_utils

django-tagulous = "*"  # https://github.com/radiac/django-tagulous
django-admin-sortable2 = "*"  # https://github.com/jrief/django-admin-sortable2
django-axes = "*"  # https://github.com/jazzband/django-axes
requests = "*"  # https://github.com/psf/requests
docker-compose = { version = "*", optional = true }  # install via: poetry install --extras "docker"
psycopg2-binary = { version = "*", optional = true }  # install via: poetry install --extras "postgres"
pillow = "*"

[tool.poetry.extras]
docker = ["docker-compose"]
postgres = ["psycopg2-binary"]

[tool.poetry.dev-dependencies]
poetry-publish = "*"  # https://github.com/jedie/poetry-publish
python-creole = "*"  # https://github.com/jedie/python-creole
tox = "*"
pytest = "*"
pytest-randomly = "*"
pytest-cov = "*"
pytest-django = "*"
coveralls = "*"
isort = "*"
flake8 = "*"
flynt = "*"
autopep8 = "*"
pyupgrade = "*"
model_bakery = "*"  # https://github.com/model-bakers/model_bakery

[tool.poetry.scripts]
manage = "inventory_project.manage:main"
update_rst_readme = "inventory_project.publish:update_readme"
publish = "inventory_project.publish:publish"

[tool.autopep8]
# https://github.com/hhatto/autopep8#pyprojecttoml
max_line_length = 120
exclude = "*/migrations/*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
